[refactor] package구조 변경 - 기존 request,response별로 dto내부에 패키지 구조를 나눴으나 실제로는   request, response모두에서 쓰이는 클래스가 있는 등, 좋지 않은 구조로   판단되어 수정
