+++ src/main/java/flab/project/data/dto/GetProfilePageResponseDto.java
@@ -1,6 +1,6 @@
-package flab.project.data.dto.response;
+package flab.project.data.dto;
 
-import flab.project.data.dto.model.Profile;
+import flab.project.data.dto.Profile;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
