+++ src/main/java/flab/project/data/dto/User.java
@@ -1,5 +1,6 @@
-package flab.project.data.dto.model;
+package flab.project.data.dto;
 
+import flab.project.data.dto.model.BasicUser;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
